move workload_ref to own module
